ref: extract table export to its own class
