+++ src/test/java/ch/admin/bar/siard2/api/TableTester.java
@@ -20,7 +20,7 @@ public class TableTester {
     private static final File _fileSIARD_21_NEW = new File("src/test/resources/tmp/sql2008new.siard");
     private static final ConfigurationProperties _cp = new ConfigurationProperties();
     private static final File _fileLOBS_FOLDER = new File(_cp.getLobsFolder());
-    private static final File _fileOE = new File("src/test/resources/testfiles/sfdboe.siard");
+    private static final File SFDBOE_SIARD = new File("src/test/resources/testfiles/sfdboe.siard");
     private static final String _sDBNAME = "SIARD 2.1 Test Database";
     private static final String _sDATA_OWNER = "Enter AG, Rüti ZH, Switzerland";
     private static final String _sDATA_ORIGIN_TIMESPAN = "Second half of 2016";
@@ -354,20 +354,16 @@ public class TableTester {
     }
 
     @Test
-    public void testExportAsHtml() {
-        try {
-            Archive archive = ArchiveImpl.newInstance();
-            // archive.open(new File("../SiardGui/testfiles/sfdboe.siard"));
-            archive.open(_fileOE);
-            Schema schema = archive.getSchema("OE");
-            File fileTable = new File("src/test/resources/tmp/CUSTOMERS.html");
-            FileOutputStream fosTable = new FileOutputStream(fileTable);
-            schema.getTable("CUSTOMERS")
-                  .exportAsHtml(fosTable, new File("src/test/resources/tmp/lobs"));
-            fosTable.close();
-            archive.close();
-        } catch (IOException ie) {
-            fail(EU.getExceptionMessage(ie));
-        }
-    } 
+    public void testExportAsHtml() throws IOException {
+        Archive archive = ArchiveImpl.newInstance();
+        archive.open(SFDBOE_SIARD);
+        Schema schema = archive.getSchema("OE");
+        File fileTable = new File("src/test/resources/tmp/CUSTOMERS.html");
+        FileOutputStream fosTable = new FileOutputStream(fileTable);
+        schema.getTable("CUSTOMERS")
+              .exportAsHtml(fosTable, new File("src/test/resources/tmp/lobs"));
+        fosTable.close();
+        archive.close();
+    }
+
 }
